Refactor folder selection logic into folder-picker utility
